Move ESIL into its own namespace
