+++ librz/arch/esil/esil_stats.c
@@ -1,6 +1,7 @@
 // SPDX-FileCopyrightText: 2014 pancake <pancake@nopcode.org>
 // SPDX-License-Identifier: LGPL-3.0-only
 
+#include <rz_esil/rz_esil.h>
 #include "analysis_private.h"
 
 static int hook_flag_read(RzAnalysisEsil *esil, const char *flag, ut64 *num) {
