+++ librz/arch/esil/esil_trace.c
@@ -2,12 +2,11 @@
 // SPDX-FileCopyrightText: 2015-2020 rkx1209 <rkx1209dev@gmail.com>
 // SPDX-License-Identifier: LGPL-3.0-only
 
+#include <rz_esil/rz_esil.h>
 #include "analysis_private.h"
 
 #define CMP_REG_CHANGE(x, y) ((x) - ((RzAnalysisEsilRegChange *)(y))->idx)
 #define CMP_MEM_CHANGE(x, y) ((x) - ((RzAnalysisEsilMemChange *)(y))->idx)
-
-#define ESILISTATE esil->analysis->esilinterstate
 
 // IL trace wrapper of esil
 static inline bool esil_add_mem_trace(RzAnalysisEsilTrace *etrace, RzILTraceMemOp *mem) {
@@ -122,10 +121,10 @@ static int trace_hook_reg_read(RzAnalysisEsil *esil, const char *name, ut64 *res
 		// RZ_LOG_WARN("Register not found in profile\n");
 		return 0;
 	}
-	if (ESILISTATE->callbacks.hook_reg_read) {
+	if (esil->analysis->esilinterstate->callbacks.hook_reg_read) {
 		RzAnalysisEsilCallbacks cbs = esil->cb;
-		esil->cb = ESILISTATE->callbacks;
-		ret = ESILISTATE->callbacks.hook_reg_read(esil, name, res, size);
+		esil->cb = esil->analysis->esilinterstate->callbacks;
+		ret = esil->analysis->esilinterstate->callbacks.hook_reg_read(esil, name, res, size);
 		esil->cb = cbs;
 	}
 	if (!ret && esil->cb.reg_read) {
@@ -166,10 +165,10 @@ static int trace_hook_reg_write(RzAnalysisEsil *esil, const char *name, ut64 *va
 
 	RzRegItem *ri = rz_reg_get(esil->analysis->reg, name, -1);
 	add_reg_change(esil->trace, esil->trace->idx + 1, ri, *val);
-	if (ESILISTATE->callbacks.hook_reg_write) {
+	if (esil->analysis->esilinterstate->callbacks.hook_reg_write) {
 		RzAnalysisEsilCallbacks cbs = esil->cb;
-		esil->cb = ESILISTATE->callbacks;
-		ret = ESILISTATE->callbacks.hook_reg_write(esil, name, val);
+		esil->cb = esil->analysis->esilinterstate->callbacks;
+		ret = esil->analysis->esilinterstate->callbacks.hook_reg_write(esil, name, val);
 		esil->cb = cbs;
 	}
 	return ret;
@@ -202,10 +201,10 @@ static int trace_hook_mem_read(RzAnalysisEsil *esil, ut64 addr, ut8 *buf, int le
 		RZ_FREE(mem_read);
 	}
 
-	if (ESILISTATE->callbacks.hook_mem_read) {
+	if (esil->analysis->esilinterstate->callbacks.hook_mem_read) {
 		RzAnalysisEsilCallbacks cbs = esil->cb;
-		esil->cb = ESILISTATE->callbacks;
-		ret = ESILISTATE->callbacks.hook_mem_read(esil, addr, buf, len);
+		esil->cb = esil->analysis->esilinterstate->callbacks;
+		ret = esil->analysis->esilinterstate->callbacks.hook_mem_read(esil, addr, buf, len);
 		esil->cb = cbs;
 	}
 	return ret;
@@ -240,10 +239,10 @@ static int trace_hook_mem_write(RzAnalysisEsil *esil, ut64 addr, const ut8 *buf,
 		add_mem_change(esil->trace, esil->trace->idx + 1, addr + i, buf[i]);
 	}
 
-	if (ESILISTATE->callbacks.hook_mem_write) {
+	if (esil->analysis->esilinterstate->callbacks.hook_mem_write) {
 		RzAnalysisEsilCallbacks cbs = esil->cb;
-		esil->cb = ESILISTATE->callbacks;
-		ret = ESILISTATE->callbacks.hook_mem_write(esil, addr, buf, len);
+		esil->cb = esil->analysis->esilinterstate->callbacks;
+		ret = esil->analysis->esilinterstate->callbacks.hook_mem_write(esil, addr, buf, len);
 		esil->cb = cbs;
 	}
 	return ret;
@@ -283,11 +282,11 @@ RZ_API void rz_analysis_esil_trace_op(RzAnalysisEsil *esil, RZ_NONNULL RzAnalysi
 	}
 	/* save old callbacks */
 	int esil_verbose = esil->verbose;
-	if (ESILISTATE->callbacks_set) {
+	if (esil->analysis->esilinterstate->callbacks_set) {
 		RZ_LOG_ERROR("esil: Cannot call recursively\n");
 	}
-	ESILISTATE->callbacks = esil->cb;
-	ESILISTATE->callbacks_set = true;
+	esil->analysis->esilinterstate->callbacks = esil->cb;
+	esil->analysis->esilinterstate->callbacks_set = true;
 
 	RzILTraceInstruction *instruction = rz_analysis_il_trace_instruction_new(op->addr);
 	rz_pvector_push(esil->trace->instructions, instruction);
@@ -305,8 +304,8 @@ RZ_API void rz_analysis_esil_trace_op(RzAnalysisEsil *esil, RZ_NONNULL RzAnalysi
 	rz_analysis_esil_parse(esil, expr);
 	rz_analysis_esil_stack_free(esil);
 	/* restore hooks */
-	esil->cb = ESILISTATE->callbacks;
-	ESILISTATE->callbacks_set = false;
+	esil->cb = esil->analysis->esilinterstate->callbacks;
+	esil->analysis->esilinterstate->callbacks_set = false;
 	esil->verbose = esil_verbose;
 	/* increment idx */
 	esil->trace->idx++;
